fix(protocol,server): 状态属性支持负值范围（声望 -100）；治理关单置 done 与重审取最新提案
- AttributeValue::Static u8 → i32：set() 对 Static 原样存储（边界钳制统一由 StatusComponent::apply_change 按 metadata 执行），u8 变体（先天属性）维持 0..255；修复声望（min=-100）经 apply_change 正确算出负值后被 set() 硬钳回 0 的链路，附回归测试
- WorldState wire 不受影响（self_state.attributes 本就是 HashMap<String,i32>），无需协议版本 bump
- 治理 M2：admin approve/reject 关单补 stage='done'（对齐 close_stale_groups 先例）——修复同类新提议因 reopen CASE 仅认 done 而追加进已关闭 group 永不重审的断链
- 治理 M3：get_group_proposal_ids 改 DESC——引擎按 first() 取样代表提案，重开场景审触发重开的新证据而非上一轮已审过的最旧提案

+++ crates/server/src/game_data/types/attribute_component.rs
@@ -204,7 +204,7 @@ impl AttributeComponent {
                             rng.random_range(min..=max)
                         })
                         .unwrap_or(10);
-                    AttributeValue::Static { value: val }
+                    AttributeValue::Static { value: val as i32 }
                 }
                 AttributeType::DailyRandom => {
                     // 每日随机属性（福缘）：每游戏日随机刷新
